+++ artifastring/violin_string.h
@@ -46,7 +46,7 @@ public:
      * @param[in] ratio_from_nut Measured as a fraction of string
      * length.
      */
-    void finger(const double ratio_from_nut);
+    void finger(const float ratio_from_nut);
 
     /** \brief Plucks a string.
      *
@@ -58,7 +58,7 @@ public:
      * @param[in] pluck_force Measured in 0.0 to 1.0 (ARBITRARY).
      * @todo Re-think the pluck force paramater + constants
      */
-    void pluck(const double ratio_from_bridge, const double pluck_force);
+    void pluck(const float ratio_from_bridge, const float pluck_force);
 
     /** \brief Sets the bow's action.
      *
@@ -68,8 +68,8 @@ public:
      * @param[in] bow_force Measured in Newtons.
      * @param[in] bow_velocity Measured in meters / second.
      */
-    void bow(const double bow_ratio_from_bridge,
-             const double bow_force, const double bow_velocity);
+    void bow(const float bow_ratio_from_bridge,
+             const float bow_force, const float bow_velocity);
 
 
     /** \brief Advances time and writes data to a buffer.
@@ -78,7 +78,7 @@ public:
      * @warning You are responsible for allocating sufficient
      * memory for the buffer.
      */
-    void fill_buffer(double* buffer, const int num_samples);
+    void fill_buffer(float* buffer, const int num_samples);
 
 
     /** \brief Returns the string physical constants.
@@ -98,83 +98,83 @@ public:
 
 protected:
 #ifdef DEBUG
-    double time_seconds;
+    float time_seconds;
 #endif
 
     // physical constants
     String_Physical pc;
 
     // optimized shared function
-    inline void calculate_eigens(double eigens[], const double position);
+    inline void calculate_eigens(float eigens[], const float position);
 
 
     // calculated from physical constants
-    double pc_c_sqrt_two_div_l;
-    double pc_c_bridge_forces[MODES];
-    double X1[MODES], X2[MODES], X3[MODES];
-    double Y1[MODES], Y2[MODES], Y3[MODES];
+    float pc_c_sqrt_two_div_l;
+    float pc_c_bridge_forces[MODES];
+    float X1[MODES], X2[MODES], X3[MODES];
+    float Y1[MODES], Y2[MODES], Y3[MODES];
     bool recache_pc_c;
     void cache_pc_c();
 
 
     // violinist physical actions
-    double vpa_finger_x1;      // measured in ratio from bridge
-    double vpa_pluck_force;    // measured in ARBITRARY
-    double vpa_bow_x0;         // measured in ratio from bridge
-    double vpa_bow_force;      // measured in newtons
-    double vpa_bow_velocity;   // measured in meters per second
+    float vpa_finger_x1;      // measured in ratio from bridge
+    float vpa_pluck_force;    // measured in ARBITRARY
+    float vpa_bow_x0;         // measured in ratio from bridge
+    float vpa_bow_force;      // measured in newtons
+    float vpa_bow_velocity;   // measured in meters per second
 
     // calculated from violinist physical actions (en masse)
-    double vpa_c_C11;
-    double vpa_c_C12;
-    double vpa_c_C01;
-    double vpa_c_C02;
+    float vpa_c_C11;
+    float vpa_c_C12;
+    float vpa_c_C01;
+    float vpa_c_C02;
     bool recache_vpa_c;
     void cache_vpa_c();
 
     // calculated from violinist physical actions (independently)
-    double vpa_c_finger_eigens[MODES];
-    double vpa_c_bow_eigens[MODES];
+    float vpa_c_finger_eigens[MODES];
+    float vpa_c_bow_eigens[MODES];
 
 
     // calculated (almost) all the time
-    double tick();
+    float tick();
 
     // do we need to calculate anything?
     bool m_active;
-    inline void check_active(const double bridge_force);
+    inline void check_active(const float bridge_force);
 
     // "historical" calculations
     // modal arrays: m_a m_adot m_a_h m_adot_h
-    double m_a_h[MODES];    // "historical" displacement
-    double m_adot_h[MODES]; // "historical" velocity
+    float m_a_h[MODES];    // "historical" displacement
+    float m_adot_h[MODES]; // "historical" velocity
     inline void compute_hist_modes();
     // calculates m_y1_h
-    double m_y1_h;
+    float m_y1_h;
     inline void compute_hist_finger();
     // calculates m_y0dot_h
-    double m_y0dot_h;
+    float m_y0dot_h;
     inline void compute_hist_bow();
     // F0 and F1
-    double m_string_excitation;
-    double m_finger_dampening;
+    float m_string_excitation;
+    float m_finger_dampening;
 
     // how should we differ from the "historical" values?
-    inline double compute_finger();
-    inline double compute_pluck();
+    inline float compute_finger();
+    inline float compute_pluck();
     bool m_bow_slipping;
-    inline double compute_bow();
+    inline float compute_bow();
 
     // values for the "next step"
-    double m_a[MODES];      // displacement
-    double m_adot[MODES];   // velocity
+    float m_a[MODES];      // displacement
+    float m_adot[MODES];   // velocity
     inline void apply_forces();
 
     // actual output of model
-    inline double compute_bridge_force();
+    inline float compute_bridge_force();
 
     // floating-point equality
-    inline bool almostEquals(double one, double two);
+    inline bool almostEquals(float one, float two);
 };
 #endif
